fix: KI-Review in Wochenplan-Seite integrieren statt separate Route
Das Wochen-Review wird jetzt direkt unter dem Wochenplan angezeigt,
synchronisiert mit der bestehenden Wochennavigation. Standalone-Route
/weekly-review leitet auf /plan um. Sidebar-Eintrag entfernt.
